Fixes 692992.  Implicit closepaths not properly detected.
GL/2 closepaths, implicitly detected when the first point of a subpath
and last are equal, were not recognized in some cases because we were
doing a comparison of transformed floating point values, the
comparison is now done correctly in fixed point.

Many regression test files have progressions where the first and last
point of a path now have a proper join where one did not exist before.