Handle the remaining CIDFonts that FAPI was missing.
This causes a number of cluster file differences, almost all are the usual
single pixel differences. There are few cases slightly better, and a few
slightly worse.

One file is a "regression" (depending on your point of view) which is badq.pdf
where the two contours making up the lower case "q" both go in the same
direction, so non-zero winding for filling results in the "loop" in the "q"
being completely filled.

The font is to blame here, but I will continue to look into a solution for
Freetype.