Bug 696636: New scan converter.
This commit adds a new scan converter implementation designed
to address some of the problems seen with 'busy' paths in the
current implementation.

No 'active line list' is kept, instead we run through the path
generating a buffer of scanline edge intersections. This trades
repeated sorting of the active list against potentially higher
memory usage.

This code has 2 pairs of different variants. Each pair copes
with 'centre of pixel' and 'any part of a pixel' rendering
modes. One pair scan converts to rectangles, another pair to
trapezoids.

These routine does not give identical results to the old code,
partly due to (minimal) rounding issues, but more due to the
fact that the new code is only equivalent to adjust values of
0 or 0.5.

GS uses intermediate adjust values by default (such as 0.3)
when rasterising at low resolution (< 150dpi). In such cases
this new code will cover more than the old code did. It is
hoped that the impact of this is reduced due to the fact we
use freetype for character rendering now.

This new scan converter is disabled by default. To enable it,
set the scanconverter value to be non-zero. This is most easily
achieved by using -dSCANCONVERTERTYPE=1 on the command line.

Runs vastly faster on the cases highlighted in bugs 692351
and 691984.