Planar device memory setup tweaks.
When setting up a memory planar device, we currently build a mask
out of the plane information supplied to us about how to pack
colors into a gx_color_index. This seems like a reasonable thing to
do as we will always need to pack colors into a gx_color_index at
some point.

As part of his work on tiffsep, Michael is about to lift this
requirement though, enabling us to cope with larger numbers of
spots. The code as is fails on certain machines/compilers due to
C's undefined behaviour when shifting by more bits than are in the
variable. We spot this case explicitly, and don't bother checking
for overlap in this case.