Memento tweaks; add Memento_breakOnFree/Realloc functionality.
Memento_breakOnFree(address) will cause a breakpoint when the block
including a given address is freed. If the block is realloced (and
hence moves elsewhere) the breakpoint is not triggered until the new
block is itself freed.

Memento_breakOnRealloc(address) will cause a breakpoint when the block
including a given address is freed or realloced.