Fix for issues in use of /SeparationOrder and /SeparationColorNames
Several issues and quite a bit of confusion in the code with respect
to this option.  I believe this should clear some things up.
Documentation still needs to be updated as to how this option functions
and what devices it actually works with.  Note that
SeparationOrder and SeparationColorNames really only works for the
tiffsep device.  The psdcmyk device was never really set up for use
with this option.  Not sure if we want to add it.  Also, I discovered
that with the disabling of compressed color encoding, the tiffsep1
device renders incorrectly.  I had not converted this device to planar
as I had thought that it performed halftoning during rendering.  I
did not realize it was rendering 8 bit data and then doing
a thresholding operation.  We may want to just move this to a planar
based device.  In that case, we could use the fast planar halftoning.

Note that with this fix, the device will only create output for the
colorants listed in /SeparationOrder.  The psdcmyk device was not
making use of the /SeparationOrder information properly.  It is now
which makes for some different renderings in the ps3cet/29-07*.ps test
files which exercise  /SeparationOrder changes.  In such a case, the
device will not output any missing colorants, which previously
it was doing.