Changes to support the new fill code which allows filling when the pen is up.
With the graphics library implementation of the hpgl/2 style fill code
it is necessary to avoid unnecessary consecutive moveto's which were
harmless in the previous code.  In particular, closepath works
differently if there are 2 or more consecutive moveto's.  Using normal
PostScript style paths the closepath will return to the last moveto in
the path, the new code will convert a consecutive moveto to "gapto"
and retain the first moveto and closepath will return to the original
moveto.