pdfwrite - address problems with stringwidth and composite fonts
Bug #693059 "Asian fonts placed badly with pdfwrite output"

The problem here was caused by the way that pdf_process_string deals with 'widths'
when responding to a strginwidth operation versus a regular show.

When handling stringwidth the width of each glyph is added to the enumerator
'returned.total_width'. However, when handling a regular show, the width of
the glyphs are simply stored directly into 'returned.total_width'.

To handle the latter case the higher level code maintains its own total width
but that was getting confused when a stringwidth was used.

Ideally we would fix pdf_process_string to behave the same whether in response
to a stringwidht or a show, but that routine is called from other places and
making it coherent looks difficult. Instead we now process stringwidth
differently form show, and don't accuulate wodths when we have a steingwidth
operation in the higher level composite font code routines.

No differences expected