pdfwrite - fix /OtherSubr stripping with Multiple Master fonts
Commit:
9f656faebac150c6aceca16c5ef6a78b92b6da43

exposed a problem with this code, when we are stripping /OtherSubrs from MM
fonts, the new CharString could be larger than the original (this can't
happen for non-MM fonts). Because we only allocated a buffer big enough to
hold the original (unstripped) data, we ended up with a buffer overflow.

The code now calls the 'strip' routine twice, the first time to calculate
how big a buffer will be required, the second time to actually do the work.

This will fix the crash in test-setweightvector.ps introduced by the
earlier commit