Add new -debug=validate-chunks debug flag.
This causes a validation to be run after every operator is interpreted.

Also, update the chunk validation code to give more informative error
messages when problems are found. (For instance, if a reference is
found to be duff, it's often the source of that reference you are
interested in, not the reference itself).