Bug 693070: psdrgb wasn't giving corrupted output.
There were actually 2 bugs in here. The first one was spotted and
fixed by James Cloos - many thanks; the arguments to memcpy were
reversed.

The second one was to do with the fact that postscript operation
now thinks it has 14 colors. When we encode/decode, we were packing
the colors into a word in such a way that we were losing the interesting
ones off the top. Reverse the order of packing, and all is well.

This still leaves an "interesting" fact about this device; we loop:

  for (chan_idx = 0; chan_idx < num_comp; chan_idx++)
    ...
    for (j = 0; j < xc->height; j++)
    ...
       get_bits_rectangle

So, in the case where we have n components, and more than 1 band
we'll be rendering the entire file n times - despite the fact that
we request all the colors each time!