pdfwrite - memory problems, don't free dictionary parameter lists too early
Bug #693069 "Regression: lockup with pdfwrite"

The image dictioanry parameter lists for distillerparams were being freed
after they were added to the 'parent' parameter list. If the same memory
should then be reused for another parameter, the parent list would end up
with corrupted linked list pointers leading to a variety of problems
includign seg faults and loops.

Although this appears as a regression its not really. The existing code only
worked by fluke, all that the offending commit did was change he memory
layout, triggering the problem.

My initial feeling was that the c_param_write code should 'deep copy' dict
parameters, just as it does for other complex objects. However reviewing the
existing usage makes it clear that this is *not* what is done at present.

Instead the list is simply inserted, and only freed when the parent list is
released. This seems to me to be poor design, because other complex types
are handled differntly, but that's the way it is. Copying the dictionary
would mean reviewing all the places that use it and modifying them. It
makes more sense to alter the pdfwrite code so that it doesn't release and
free the dictioanry list.

Took the opportunity to refactor a few places in the code which use massive
'if' clauses which call code. This construct is confusing and difficult to
debug.

No expected differences