pdfwrite - properly mark cos_stream as closed when closing associated stream
Bug #6930 "Regression: /rangecheck in --pdfmark-- starting with 25b6a2cdeb5a90bb7803958f2bf15b618fbcf120"

Commit 25b6a2cdeb5a90bb7803958f2bf15b618fbcf120 was a little too
enthusiastic. In addition to properly closing the streams, I inadvertently
removed the code which marks the pdfwrite cos_stream object as closed when
we close the stream.

Ordinarily this doesn't matter, but the pdfmark code checks to maek sure the
stream associated with a named object is closed when it is used, and the
removal of the code left it apparently open, causing the error.

No differences expected