Fix for broken AA support for devices that support the devn color type
This fix involved the addition of a copy_alpha_hl_color for passing along the devn color
value when doing the copy_alpha procedure.  This required support through the clist,
special handing in the pdf14 device and a default procedure for the operation.
The only devices that should be affected are tiffsep and psdcmyk.  Support for 16bit psd
devices may have issues and I will go back to check on this later as a customer is waiting
for this for 8 bit tiffsep.