pdfwrite - Make sure TT fonts in PDF files are named using the PDF font name
Bug #693086

The problem was that the TrueType code used by the PDF interpreter uses the
font name from the TrueType font, not the name from the PDF file. If two
subsets are present then pdfwrite is unable to distinguish them if the
Encodings are apparently the same (as they are here) as the fonts have have
the same name internally, even though they are different in the PDF.

TrueType CIDFonts already take the font naem form the PDF file and apply that
so here we simply do the same for regular TrueType fonts.

This causes a small difference in Bug689073.pdf, apparently becuase it
contains two fonts called Arial,Bold