pdfwrite - don't write CIDSet for subset CIDFonts when emitting PDF/A-2
The specification for CIDSet has changed for PDF/A-2 and I am currently
unable to create a CIDSet which the Acrobat pre-flught tool will validate.

Since PDF/A-2 now makes CIDSet optional, this commit omits it when we are
producing PDF/A-2, so that we can create a valid file.

No differences expected