pdfwrite - fix handling of transparent images from PDF interpreter
The PDF interperter can insert extra groups when dealing with transparent
images, which we don't want to emit in the output PDF. These are marked
by a 'image_with_SMask' flag.

Previously we stored a boolean in the device structure, and tested that, if
it was true we skipped the begin and transparency group operations.

It turns out that it is possible to get a new group *inside* the generated
group we are skipping, if that happens then the second group would reset
the flag. This led to us skipping the 'begin' but not the 'end' with predictabley
bad consequences.

In this commit we alter the boolena to an  unsigned int, and treat it as a
series of bits. We flip the bits based on the 'FormDepth' which we track
separately. This allows up to 32 levels of nesting which ought to be enough
to deal even with Cairo files.

Expected differences
Bug689897.pdf exhibits small differences, not visually apparent. I believe this
is a progression but its so slight as to be difficult to tell.