Bug 692365: Don't derive metrics for CIDFontType 2.
pdfwrite uses zchar42_set_cache() to get the glyph metrics for a glyph from
CIDFontType 2 font. It turns out, that function contained the same faking
of vertical metrics as the FAPI code originally had. That's a problem because
a PS font without vertical metrics should be treated as horizontal, regardless
of the wmode value.

We *only* want to derive vertical metrics for TTF fonts read from disk being
used to emulate a missing CIDFont with wmode 1.

No cluster differences.