Add 3:2 and 3:4 downscaling to tiffsep/psd/downscaler.
Currently the downscaler can only downscale in integer increments. To
accomodate a potential need to efficiently scale 1200 -> 800 and
600 -> 800 dpi, we introduce new functionality to allow 3:2 and 3:4
scaling modes.

We shoehorn these into the existing scaler system by using DownScaleFactor
settings of 32 and 34 respectively; any other DownScaleFactor > 8 will
give a rangecheck error.

This has required some changes within the downscaler code itself, and
will require more changes in any device that wants to use these. Currently
the cores are only provided in the planar modes; hence tiffsep and psd are
the only devices that have been updated to work with this.