Bug 693064: raise maximum possible GS_CLIENT_COLOR_MAX_COMPONENTS to 32
By default we support a maximum of 14 components. Supposedly this can be
increased by raising GS_CLIENT_COLOR_MAX_COMPONENTS to a larger number
on startup, but this starts to cause problems in various places throughout
the code.

The first such place is in the bpc_to_depth function (found in gdevdevn
and various other places), where the calculation goes wrong for anything
above 31 components at 8 bpc. We fix that here.

This allows us to get to 32 components. To raise it above 32 presents
problems on most architectures as the code assumes elsewhere that we
can use a bitmask to represent which components are present.

We may be able to tweak the code to use a uint64_t instead, in which
case we can probably get to 64 components; is that high enough?