pdfwrite - record glyph vertical metrics evne when the position vector has a vertical shift
Bug #692365

The store_glyph_width routine returns >0 if there is any vertical shift in
addition to a width, this is defined as 'not safe to cache'. The routine
pdf_glyph_widths ORs that into its return value and passes the data back.

scan_cmap_text was checking that value before recording the width, and not
recording it if it was > 0. This is incorrect, at least when in vertical writing
mode, as this is the only place that we record the vertical metrics for use
in the W2 array. If we don't record them here this information is lost. Also
we don't need to worry about the vertical movement here as that is dealt with
elswewhere by modifying the text matrix.

This was leading to pdfwrite emitting a /W2 entry with all 0 values [0 0 0]
which caused significant misplacement of glyphs.

Expected Differences
Bug692576.ps exhibits a progression