pdfwrite - store widths for regular fonts when there is a vertical displacement.
Bug #691369 "Widths array in font dictionary contains incorrect values for rotated text"

Similar to bug #692365 but this time for regular fonts not CIDFonts. If there
is any vertcial displacement of the glyph then we did not store teh glyph
width, leading to it being emitted as 0.

Fix is very similar also.

No differences expected.