Fix in XPS interpreter to ensures we only push the smallest transparency group needed + more
For a path that included an opacity mask, the XPS interpreter was pushing a soft mask
that included the entire page (it actually just looked at the current clip path). This
fix incorporates the path that we are filling in the computation of the transparency group
size.  Also, there were issues with the softmasks in general that were causing them to be
double applied.  Fixes 693042 and also provides improvements for 690951.  Also this should fix 693137 and 692892