Use .setuserparams2 instead of .setuserparams when setting the ICC user params
Thanks to Chris for pointing this out.  If we use .setuserparams then in gs_lev2.ps the
various user params such as /DefaultCMYKProfile are set.  This sets the profile
in the icc manager properly to what ever was set on the command line.
Unfortunately, the currentuserparams dictionary (or what ever object it is)
that is maintained by the interpreter is not updated with the same value, but
instead it has an old default value that it obtained from an earlier call of
.currentuserparams.   During a vm_reclaim,  context_state_store stores the
outdated userparam information and then context_state_load  does a call to
set_user_params with this outdated userparams information.  setuserparams2
takes care of this by getting the interpreters list in sync at the time we do
the setting of the parameter.  Fixes bug 693159