Bug 693176: write transparent type2 pattern color to clist
When using a clist, ensure that the all the color space data for the
pattern gets written to the clist, *and* that the clist correctly
records all the relevant transparency data.

I'm at a loss to explain how COMPILE_INITS influenced this, but this
change brings the clist/non-clist output and code in line with each
other.

No cluster differences.