Remove an additional, incorrect, search for resources, intended for OpenVMS only.
ResourceForAll used a 2nd template for searching file-based resources,
    GenericResourceDir + Category
      + .file_name_separator + originaltemplate + .file_name_separator + "*"
in addition to the standard
    GenericResourceDir + Category + .file_name_separator + originaltemplate
This was documented as needed for OpenVMS, but it results in
    dev:[genericresourcedir.Category]originaltemplate]* ("]" twice)
which is definitely invalid. It is also unnecessary, the "standard" one is OK.

For other platforms, this 2nd form may or may not be syntacticly valid ("*?",
coming in form the original template, in the "directory" part), but it's
definitely not useful.

Of course, it remains the question what was the intent of the person who added
these lines (in case we should keep the code but fix it). I considered:

- "dev:[genericresourcedir.Category.*]template", but this searches one
  directory level too deep.
- "dev:[genericresourcedir.Category]template;*", but this would load all
  versions of a same file. Definitely not what we want, because we may end
  with an obsolete (old) copy of the resource.
- "dev:[genericresourcedir.Category]template.*", but this is unnecessary. And
  maybe more important is that the exact template used is the responsability
  of the caller; I see no reason for ResourceForAll to modify it.

The most precise documentation on OpenVMS path syntax I found is "Guide to
OpenVMS File Applications"
<http://h71000.www7.hp.com/doc/731final/documentation/pdf/ovms_731_file_app.pdf>
available from <http://h71000.www7.hp.com/doc/os83_index.html>.
I also asked on comp.os.vms, and the response I got confirmed the form used is
invalid.