Bug 693155: 'resourceforall' returns truncated names for /CMap resources.
(A)
Comments say that gs_init.ps::.generate_dir_list_templates_with_length returns
the legth of the "directory" part, but if actually returned the length of
"[base] directory" + "template" - 1.

(B)
gs_res.ps::ResourceForAll passes part of the real "directory" (the "Category/")
as part of the "template" argument of .generate_dir_list_templates_with_length,
but failed to take into consideration the length of this string when cutting
"directory" from the full filespec in order to remain with the resource name.

Note that .generate_dir_list_templates_with_length does not check if its
"template" argument contains directories too so cannot adjust the returned
length. The code as written intends to deal with resource names that include
.file_name_separator (see discussion on Bug 688737 "'resourceforall' truncates
names of file-based resources"); such a resource name would be
"Times/Cyrillic".