make the SEAC detector and type 1 -> type 2 charstring converter tolerate broken fonts
Bug #693170 "ps2pdf crashes on a given file"

The fonts in the PostScript file are invalid in that some of the glyphs do not
end with an 'endchar' or other valid termination operator. This was causing
a buffer overrun which eventually could cause access to invalid memory and a crash.

Added limit checking where it was missing and choose to return an 'unknown'
error so that code which is able to handle this can detect the condition. Code
which is not expecting this contdition will continue to signal an error.

No differences expected