Bug 692684: stack variable corruption
In alloc_restore_step_in(), we have to make sure that, whenever we back up
a memory space to the previous save level, we hold onto the gc_status.psignal
pointer, and put it into the restored memory space. In case the "save"
happened during an earlier call to interp() and co.

No cluster differences.