Bug 693166: Optimise interpolation
When interpolating, ghostscript pays no heed to the clipping rectangle.
Hence if we scale (say) a 256x256 image up to (say)17067x17067, even
though only a small portion of the scaled up image is actually visible
we scale the whole lot only to throw away 90%+ of it.

To fix this, we have to extend the capabilities of the interpolation
code.

The existing code already copes with only being given data for a
subsection of the image (for when we split images in the clist, I
guess). This rectangle is referred to in the code as being 'the
subrectangle we are rendering', when it's actually 'the subtrectangle
we are being given data for'. We update the description to be more
accurate.

We introduce a new rectangle, 'the render rectangle' to indicate the
subrectangle that we are actually rendering - this will always be a
subset of the data rectangle.

If we are given a clipping rectangle, we read the outer bbox from it,
and map this back into the source space of the image; we intersect this
with the data rectangle to get the render rectangle.

We update the scaling stream filter to set an 'Active' flag to say
whether we are inside the render rectangle or not. If not, we can
safely skip lines in their totality. By default we leave this set to
1, so that any scaling cores that aren't updated to know about this
will perform in the old way.

We update the scaling code to make use of the Active flag; whole lines
are skipped if we aren't in the active region, and if we are, we skip
prefixes/suffixes of unused pixels.

We update the scaling cores themselves to avoid calculating values
outside the active regions.

Note that for simplicity we still allocate space as if we were
accessing the whole line. We still calculate contributions for the whole
of the images; to do otherwise would require significant changes to
the weight generation code, and this isn't a huge consumer of time.