pdfwrite - code cleanup
Deprecated the undocumented 'MaxViewerMemorySize' parameter whcih wasn't in
fact being used anywhere useful.

Deprecated the 'pdf_write_and_free_all_resoruces' function which is not used
after deprecating MaxViewerMemorySize.

Deprecated the function 'pdf_glyph_origin' as it is unused

made 'write-font_resources' non-static so it can be called from pdf_close

Deprecated pdf_close_text_document and moved the code (which actually writes
fonts and font descriptors) into pdf_close so that its easier to follow the
control flow. Also cleaned up the code so that it isn't one huge 'if' test
but each function call is separated out for easier debugging.

Made a number of '#if 0' blocks into '#ifdef DEPRECATED_906'. All blocks of
code marked this way will be removed after the release of 9.06.

Removed a number of cases where we had #define labels with a bug number, these
now use the #ifdef DEPRECATED_906 lable or are simply removed and the code
compiled in.

Simlarly with '#define RIGHT_SBW' and '#if !RIGHT_SBW'

Updated the ps2ps2.htm document to remove the reference to MaxViewerMemorySize.

No differences expected.