ps2write - create document %%BoundingBox from teh individual page Bounding boxes
Bug #693181

The document level BoundingBox is supposed to be the intersection of the
boundingbox of all the pages, whereas in fact it is the device media size.

Now we create the document BoundingBox by taking largest dimensions from all
the pages in the output.

NB the BoundingBox for each page is given from the media request, so this is
not a true BoundingBox anyway, but at least it is better then before.