Bug 689376: fix //name handling on command prompt.
We allocate %statementedit in the system memory. When a is read from
the command prompt we use gs_scan_token() to determine, whether the
line has a complete token. If the line has immediately evaluated names,
gs_scan_token() fails. This is interpreted as the end of the input.
If the line buffer has a complete token, the buffer is rescanned
and everything is fine. Otherwise, scanning a prematurely closed buffer
causes an error.

All we need to do to fix this is to force local memory allocation
instead of inheriting system memory from %statementedit. The scanned
token is discarded after determining whether the input line is
complete.