Add in support for copy_planes into the overprint device
The copy_planes procedure is used during the default copy_alpha_hl_color and if we
are going out to the overprint device, we need to make sure to copy only those
planes that are specified in the drawn components vector.   We also needed to
distinguish from this case and a copy_planes coming from the pattern tiling code.
This was done by adding in a copy_alpha_hl_color proc for the overprint device that
sets a flag and calls the default copy_alpha_hl_color so that later when we get
back to the overprint device's copy_planes proc we will know to do the over print
rather than just pass things along to the target device.  Fixes Bug 693184.