Bug 693188: don't skip glyphs with degenerate matrices
In commit 752397ab, the fix for Bug 692263 was to skip over glyphs when the
matrix in force for the glyph had a zero scale in either dimension.

It seems that this approach had potentially unfavourable side effects in the
glyph cache.

So, when we encounter a degenerate matrix rendering a glyph, we now force a
minimal, non-zero scale before attempting to render the glyph. This placates
the glyph cache.

Cluster testing shows (tiny) progressions in Bug689006.pdf, Bug690179.pdf,
and Bug692634.ps, and pixel differences, not identifiable as progressions
or regressions in Bug690497.pdf and 12-07C.PS.