Fix for bug in the named color example support
When the example in gs\toolbin\color\named_color is run we get a segv.  It was due to
the name information not getting set up in the nameed color structure that the example uses.
It essentially uses an ICC for with the named spot color data packed into the buffer.