ps2write - Always ascii hex encode the eexec portion of type 1 fonts
Bug #693197 "ps2write outputs Type 1 fonts in pfb format"

Technically it is possible for the eexec portion of a type 1 font to be in
binary or ascii hex. Because we always ascii85 encode binary data if the
device does not support it, we left eexec portion as binary.

However it seems that at least some printers (Kyocera, possibly others) are
unable to deal with this.

This patch means that we alays use ascii hex data for the eexec encrypted
portion of a type 1 font. We did consider making this a command line option
but were unable to find any other tools which emitted binary here so opted
simply to conform.

All the investigation and the original patch (see bug report) by Chris Liddell


No differences expected