Rework colors_used to be a color_usage bitfield.
Previously, the clist would collate the colors used in a band/page
by ORing together any color indexs uses into a single gx_color_index.
This relies on the gx_color_index being able to represent the whole
depth.

This is dodgy with compressed encoding, and fails entirely with the
new planar based tiffsep/psdcmyk and high level color stuff, as the
total depth can far exceed the number of bits available in a
gx_color_index.

The fix here is to change to using a bitfield (gx_color_usage_bits)
for this record; this allows us to have up to 64 colorants with a
standard build.

The code here is still imperfect for all the reasons listed within
the original code (only works for subtractive spaces, can falsely
detect 'no colors used', etc), but it is at least consistently
imperfect now.