Fix so that when we are doing the sep devices with spots the blend spaces for transparency remain separated
This fixes bug 693144 first part which is the segv for Bug6992217.pdf which likely existed prior to the
planar change for the sep devices.  The issue with 12-07C.PS is different.  In any event, the
issue was that there was a fill with a pattern in a softmask and there was an RGB transparency group
within the pattern.  With the separation devices, the blending space is maintained as a subtractive
space to allow blending of all the spot colorants which means that the parent group is always maintained.
There is a discussion about this in the PDF specification with respect to spot colors and transparency
blending.  In any event, the current code has some confusion since the decision as to if we should use
the group color space was based upon the device name (e.g. pdf14cmykspot) which became a problem when we
had the pdf14clistgray device when the pattern was a clist.